[refactor] 클래스명, 구조 변경 #221
