+++ src/main/java/com/triptune/schedule/controller/TravelRouteController.java
@@ -2,7 +2,7 @@ package com.triptune.schedule.controller;
 
 import com.triptune.schedule.dto.request.RouteCreateRequest;
 import com.triptune.schedule.dto.response.RouteResponse;
-import com.triptune.schedule.service.RouteService;
+import com.triptune.schedule.service.TravelRouteService;
 import com.triptune.global.aop.AttendeeCheck;
 import com.triptune.global.response.ApiResponse;
 import com.triptune.global.response.pagination.ApiPageResponse;
@@ -18,16 +18,16 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/schedules/{scheduleId}")
 @RequiredArgsConstructor
 @Tag(name = "Schedule - Route", description = "일정 만들기 중 여행 루트 관련 API")
-public class RouteController {
+public class TravelRouteController {
 
-    private final RouteService routeService;
+    private final TravelRouteService travelRouteService;
 
     @AttendeeCheck
     @GetMapping("/routes")
     @Operation(summary = "여행 루트 조회", description = "저장되어 있는 여행 루트를 조회한다.")
     public ApiPageResponse<RouteResponse> getTravelRoutes(@PathVariable(name = "scheduleId") Long scheduleId,
                                                           @RequestParam(name = "page") int page){
-        Page<RouteResponse> response = routeService.getTravelRoutes(scheduleId, page);
+        Page<RouteResponse> response = travelRouteService.getTravelRoutes(scheduleId, page);
 
         return ApiPageResponse.dataResponse(response);
     }
@@ -38,7 +38,7 @@ public class RouteController {
     public ApiResponse<?> createLastRoute(@AuthenticationPrincipal(expression = "memberId") Long memberId,
                                           @PathVariable(name = "scheduleId") Long scheduleId,
                                           @Valid @RequestBody RouteCreateRequest routeCreateRequest){
-        routeService.createLastRoute(scheduleId, memberId, routeCreateRequest);
+        travelRouteService.createLastRoute(scheduleId, memberId, routeCreateRequest);
         return ApiResponse.okResponse();
     }
